requirements.txt
geopandas
rasterio
matplotlib
numpy
pandas
seaborn
scikit-learn
streamlit
earthengine-api
notebook
ipykernel
tqdm
geemap
requests
osmnx
gdown
geedim
rasterstats
folium
streamlit-folium
plotly
pillow